Move child resource method params next to method declarations
