--- types/2022-11-15/BankAccountsResource.d.ts
@@ -1,5 +1,0 @@
-// File generated from our OpenAPI spec
-
-declare module 'stripe' {
-  namespace Stripe {}
-}
