--- types/2022-11-15/CreditNoteLineItemsResource.d.ts
@@ -1,12 +1,0 @@
-// File generated from our OpenAPI spec
-
-declare module 'stripe' {
-  namespace Stripe {
-    interface CreditNoteLineItemListParams extends PaginationParams {
-      /**
-       * Specifies which fields in the response should be expanded.
-       */
-      expand?: Array<string>;
-    }
-  }
-}
