--- types/2022-11-15/ExternalAccountsResource.d.ts
@@ -1,134 +1,0 @@
-// File generated from our OpenAPI spec
-
-declare module 'stripe' {
-  namespace Stripe {
-    interface ExternalAccountCreateParams {
-      /**
-       * Please refer to full [documentation](https://stripe.com/docs/api) instead.
-       */
-      external_account: string;
-
-      /**
-       * When set to true, or if this is the first external account added in this currency, this account becomes the default external account for its currency.
-       */
-      default_for_currency?: boolean;
-
-      /**
-       * Specifies which fields in the response should be expanded.
-       */
-      expand?: Array<string>;
-
-      /**
-       * Set of [key-value pairs](https://stripe.com/docs/api/metadata) that you can attach to an object. This can be useful for storing additional information about the object in a structured format. Individual keys can be unset by posting an empty value to them. All keys can be unset by posting an empty value to `metadata`.
-       */
-      metadata?: Stripe.MetadataParam;
-    }
-
-    interface ExternalAccountRetrieveParams {
-      /**
-       * Specifies which fields in the response should be expanded.
-       */
-      expand?: Array<string>;
-    }
-
-    interface ExternalAccountUpdateParams {
-      /**
-       * The name of the person or business that owns the bank account.
-       */
-      account_holder_name?: string;
-
-      /**
-       * The type of entity that holds the account. This can be either `individual` or `company`.
-       */
-      account_holder_type?: Stripe.Emptyable<
-        ExternalAccountUpdateParams.AccountHolderType
-      >;
-
-      /**
-       * The bank account type. This can only be `checking` or `savings` in most countries. In Japan, this can only be `futsu` or `toza`.
-       */
-      account_type?: ExternalAccountUpdateParams.AccountType;
-
-      /**
-       * City/District/Suburb/Town/Village.
-       */
-      address_city?: string;
-
-      /**
-       * Billing address country, if provided when creating card.
-       */
-      address_country?: string;
-
-      /**
-       * Address line 1 (Street address/PO Box/Company name).
-       */
-      address_line1?: string;
-
-      /**
-       * Address line 2 (Apartment/Suite/Unit/Building).
-       */
-      address_line2?: string;
-
-      /**
-       * State/County/Province/Region.
-       */
-      address_state?: string;
-
-      /**
-       * ZIP or postal code.
-       */
-      address_zip?: string;
-
-      /**
-       * When set to true, this becomes the default external account for its currency.
-       */
-      default_for_currency?: boolean;
-
-      /**
-       * Two digit number representing the card's expiration month.
-       */
-      exp_month?: string;
-
-      /**
-       * Four digit number representing the card's expiration year.
-       */
-      exp_year?: string;
-
-      /**
-       * Specifies which fields in the response should be expanded.
-       */
-      expand?: Array<string>;
-
-      /**
-       * Set of [key-value pairs](https://stripe.com/docs/api/metadata) that you can attach to an object. This can be useful for storing additional information about the object in a structured format. Individual keys can be unset by posting an empty value to them. All keys can be unset by posting an empty value to `metadata`.
-       */
-      metadata?: Stripe.Emptyable<Stripe.MetadataParam>;
-
-      /**
-       * Cardholder name.
-       */
-      name?: string;
-    }
-
-    namespace ExternalAccountUpdateParams {
-      type AccountHolderType = 'company' | 'individual';
-
-      type AccountType = 'checking' | 'futsu' | 'savings' | 'toza';
-    }
-
-    interface ExternalAccountListParams extends PaginationParams {
-      /**
-       * Specifies which fields in the response should be expanded.
-       */
-      expand?: Array<string>;
-
-      object?: ExternalAccountListParams.Object;
-    }
-
-    namespace ExternalAccountListParams {
-      type Object = 'bank_account' | 'card';
-    }
-
-    interface ExternalAccountDeleteParams {}
-  }
-}
