--- types/2022-11-15/TaxIdsResource.d.ts
@@ -1,93 +1,0 @@
-// File generated from our OpenAPI spec
-
-declare module 'stripe' {
-  namespace Stripe {
-    interface TaxIdCreateParams {
-      /**
-       * Type of the tax ID, one of `ae_trn`, `au_abn`, `au_arn`, `bg_uic`, `br_cnpj`, `br_cpf`, `ca_bn`, `ca_gst_hst`, `ca_pst_bc`, `ca_pst_mb`, `ca_pst_sk`, `ca_qst`, `ch_vat`, `cl_tin`, `eg_tin`, `es_cif`, `eu_oss_vat`, `eu_vat`, `gb_vat`, `ge_vat`, `hk_br`, `hu_tin`, `id_npwp`, `il_vat`, `in_gst`, `is_vat`, `jp_cn`, `jp_rn`, `jp_trn`, `ke_pin`, `kr_brn`, `li_uid`, `mx_rfc`, `my_frp`, `my_itn`, `my_sst`, `no_vat`, `nz_gst`, `ph_tin`, `ru_inn`, `ru_kpp`, `sa_vat`, `sg_gst`, `sg_uen`, `si_tin`, `th_vat`, `tr_tin`, `tw_vat`, `ua_vat`, `us_ein`, or `za_vat`
-       */
-      type: TaxIdCreateParams.Type;
-
-      /**
-       * Value of the tax ID.
-       */
-      value: string;
-
-      /**
-       * Specifies which fields in the response should be expanded.
-       */
-      expand?: Array<string>;
-    }
-
-    namespace TaxIdCreateParams {
-      type Type =
-        | 'ae_trn'
-        | 'au_abn'
-        | 'au_arn'
-        | 'bg_uic'
-        | 'br_cnpj'
-        | 'br_cpf'
-        | 'ca_bn'
-        | 'ca_gst_hst'
-        | 'ca_pst_bc'
-        | 'ca_pst_mb'
-        | 'ca_pst_sk'
-        | 'ca_qst'
-        | 'ch_vat'
-        | 'cl_tin'
-        | 'eg_tin'
-        | 'es_cif'
-        | 'eu_oss_vat'
-        | 'eu_vat'
-        | 'gb_vat'
-        | 'ge_vat'
-        | 'hk_br'
-        | 'hu_tin'
-        | 'id_npwp'
-        | 'il_vat'
-        | 'in_gst'
-        | 'is_vat'
-        | 'jp_cn'
-        | 'jp_rn'
-        | 'jp_trn'
-        | 'ke_pin'
-        | 'kr_brn'
-        | 'li_uid'
-        | 'mx_rfc'
-        | 'my_frp'
-        | 'my_itn'
-        | 'my_sst'
-        | 'no_vat'
-        | 'nz_gst'
-        | 'ph_tin'
-        | 'ru_inn'
-        | 'ru_kpp'
-        | 'sa_vat'
-        | 'sg_gst'
-        | 'sg_uen'
-        | 'si_tin'
-        | 'th_vat'
-        | 'tr_tin'
-        | 'tw_vat'
-        | 'ua_vat'
-        | 'us_ein'
-        | 'za_vat';
-    }
-
-    interface TaxIdRetrieveParams {
-      /**
-       * Specifies which fields in the response should be expanded.
-       */
-      expand?: Array<string>;
-    }
-
-    interface TaxIdListParams extends PaginationParams {
-      /**
-       * Specifies which fields in the response should be expanded.
-       */
-      expand?: Array<string>;
-    }
-
-    interface TaxIdDeleteParams {}
-  }
-}
