+++ types/2022-11-15/index.d.ts
@@ -11,65 +11,33 @@
 ///<reference path='./AccountLinksResource.d.ts' />
 ///<reference path='./ApplePayDomainsResource.d.ts' />
 ///<reference path='./ApplicationFeesResource.d.ts' />
-///<reference path='./Apps/SecretsResource.d.ts' />
 ///<reference path='./BalanceResource.d.ts' />
 ///<reference path='./BalanceTransactionsResource.d.ts' />
-///<reference path='./BankAccountsResource.d.ts' />
-///<reference path='./BillingPortal/ConfigurationsResource.d.ts' />
-///<reference path='./BillingPortal/SessionsResource.d.ts' />
-///<reference path='./CapabilitiesResource.d.ts' />
-///<reference path='./CardsResource.d.ts' />
-///<reference path='./CashBalancesResource.d.ts' />
 ///<reference path='./ChargesResource.d.ts' />
-///<reference path='./Checkout/SessionsResource.d.ts' />
 ///<reference path='./CountrySpecsResource.d.ts' />
 ///<reference path='./CouponsResource.d.ts' />
 ///<reference path='./CreditNotesResource.d.ts' />
-///<reference path='./CreditNoteLineItemsResource.d.ts' />
 ///<reference path='./CustomersResource.d.ts' />
-///<reference path='./CustomerBalanceTransactionsResource.d.ts' />
-///<reference path='./CustomerCashBalanceTransactionsResource.d.ts' />
 ///<reference path='./DisputesResource.d.ts' />
 ///<reference path='./EphemeralKeysResource.d.ts' />
 ///<reference path='./EventsResource.d.ts' />
 ///<reference path='./ExchangeRatesResource.d.ts' />
-///<reference path='./ExternalAccountsResource.d.ts' />
-///<reference path='./FeeRefundsResource.d.ts' />
 ///<reference path='./FilesResource.d.ts' />
 ///<reference path='./FileLinksResource.d.ts' />
-///<reference path='./FinancialConnections/AccountsResource.d.ts' />
-///<reference path='./FinancialConnections/SessionsResource.d.ts' />
-///<reference path='./Identity/VerificationReportsResource.d.ts' />
-///<reference path='./Identity/VerificationSessionsResource.d.ts' />
 ///<reference path='./InvoicesResource.d.ts' />
 ///<reference path='./InvoiceItemsResource.d.ts' />
-///<reference path='./Issuing/AuthorizationsResource.d.ts' />
-///<reference path='./Issuing/CardsResource.d.ts' />
-///<reference path='./Issuing/CardholdersResource.d.ts' />
-///<reference path='./Issuing/DisputesResource.d.ts' />
-///<reference path='./Issuing/TransactionsResource.d.ts' />
-///<reference path='./InvoiceLineItemsResource.d.ts' />
-///<reference path='./LoginLinksResource.d.ts' />
 ///<reference path='./MandatesResource.d.ts' />
 ///<reference path='./PaymentIntentsResource.d.ts' />
 ///<reference path='./PaymentLinksResource.d.ts' />
 ///<reference path='./PaymentMethodsResource.d.ts' />
-///<reference path='./CustomerSourcesResource.d.ts' />
 ///<reference path='./PayoutsResource.d.ts' />
-///<reference path='./PersonsResource.d.ts' />
 ///<reference path='./PlansResource.d.ts' />
 ///<reference path='./PricesResource.d.ts' />
 ///<reference path='./ProductsResource.d.ts' />
 ///<reference path='./PromotionCodesResource.d.ts' />
 ///<reference path='./QuotesResource.d.ts' />
-///<reference path='./Radar/EarlyFraudWarningsResource.d.ts' />
-///<reference path='./Radar/ValueListsResource.d.ts' />
-///<reference path='./Radar/ValueListItemsResource.d.ts' />
 ///<reference path='./RefundsResource.d.ts' />
-///<reference path='./Reporting/ReportRunsResource.d.ts' />
-///<reference path='./Reporting/ReportTypesResource.d.ts' />
 ///<reference path='./ReviewsResource.d.ts' />
-///<reference path='./Sigma/ScheduledQueryRunsResource.d.ts' />
 ///<reference path='./SetupAttemptsResource.d.ts' />
 ///<reference path='./SetupIntentsResource.d.ts' />
 ///<reference path='./ShippingRatesResource.d.ts' />
@@ -78,39 +46,11 @@
 ///<reference path='./SubscriptionItemsResource.d.ts' />
 ///<reference path='./SubscriptionSchedulesResource.d.ts' />
 ///<reference path='./TaxCodesResource.d.ts' />
-///<reference path='./TaxIdsResource.d.ts' />
 ///<reference path='./TaxRatesResource.d.ts' />
-///<reference path='./Terminal/ConfigurationsResource.d.ts' />
-///<reference path='./Terminal/ConnectionTokensResource.d.ts' />
-///<reference path='./Terminal/LocationsResource.d.ts' />
-///<reference path='./Terminal/ReadersResource.d.ts' />
-///<reference path='./TestHelpers/TestClocksResource.d.ts' />
 ///<reference path='./TokensResource.d.ts' />
 ///<reference path='./TopupsResource.d.ts' />
 ///<reference path='./TransfersResource.d.ts' />
-///<reference path='./TransferReversalsResource.d.ts' />
-///<reference path='./Treasury/CreditReversalsResource.d.ts' />
-///<reference path='./Treasury/DebitReversalsResource.d.ts' />
-///<reference path='./Treasury/FinancialAccountsResource.d.ts' />
-///<reference path='./Treasury/InboundTransfersResource.d.ts' />
-///<reference path='./Treasury/OutboundPaymentsResource.d.ts' />
-///<reference path='./Treasury/OutboundTransfersResource.d.ts' />
-///<reference path='./Treasury/ReceivedCreditsResource.d.ts' />
-///<reference path='./Treasury/ReceivedDebitsResource.d.ts' />
-///<reference path='./Treasury/TransactionsResource.d.ts' />
-///<reference path='./Treasury/TransactionEntriesResource.d.ts' />
-///<reference path='./UsageRecordsResource.d.ts' />
-///<reference path='./UsageRecordSummariesResource.d.ts' />
 ///<reference path='./WebhookEndpointsResource.d.ts' />
-///<reference path='./TestHelpers/CustomersResource.d.ts' />
-///<reference path='./TestHelpers/Issuing/CardsResource.d.ts' />
-///<reference path='./TestHelpers/RefundsResource.d.ts' />
-///<reference path='./TestHelpers/Terminal/ReadersResource.d.ts' />
-///<reference path='./TestHelpers/Treasury/InboundTransfersResource.d.ts' />
-///<reference path='./TestHelpers/Treasury/OutboundPaymentsResource.d.ts' />
-///<reference path='./TestHelpers/Treasury/OutboundTransfersResource.d.ts' />
-///<reference path='./TestHelpers/Treasury/ReceivedCreditsResource.d.ts' />
-///<reference path='./TestHelpers/Treasury/ReceivedDebitsResource.d.ts' />
 ///<reference path='./AccountLinks.d.ts' />
 ///<reference path='./Accounts.d.ts' />
 ///<reference path='./ApplePayDomains.d.ts' />
